feat(cli): v1.0.2 新增隧道功能和设备详情
- devices -v 显示 S5 代理、隧道状态等详细信息
- mappings 查看当前所有端口映射
- tunnel 建立端口映射（打洞）
- disconnect 断开端口映射
- 新增 RPA 步骤：condition_check, execute_repo_script, set_variables, start_bot

+++ internal/rpa/steps/change_os.go
@@ -75,7 +75,7 @@ func (s *ChangeOsAndWaitStep) sendChangeOs(deviceID int, params map[string]inter
 
 	// 构建 Changephones 请求数据
 	changeData := map[string]interface{}{
-		"deviceId": deviceID,
+		"deviceId": float64(deviceID),
 		"msisdn":   msisdn,
 		"smsc":     smsc,
 	}
